fix(query-devtools): remove solid-js import from declarations
Move public devtools types into a Solid-free module so declaration
generation does not emit a side-effect import for solid-js.
