+++ .changeset/remove-solid-declaration-import.md
@@ -1,0 +1,5 @@
+---
+'@tanstack/query-devtools': patch
+---
+
+Remove an unintended `solid-js` import from the generated declaration files.
