+++ packages/query-devtools/src/contexts/QueryDevtoolsContext.ts
@@ -1,38 +1,14 @@
 import { createContext, useContext } from 'solid-js'
-import type { Query, QueryClient, onlineManager } from '@tanstack/query-core'
+import type { QueryClient, onlineManager } from '@tanstack/query-core'
+import type { QueryDevtoolsProps } from './types'
 
-type XPosition = 'left' | 'right'
-type YPosition = 'top' | 'bottom'
-export type DevtoolsPosition = XPosition | YPosition
-export type DevtoolsButtonPosition = `${YPosition}-${XPosition}` | 'relative'
-export type Theme = 'dark' | 'light' | 'system'
-
-export interface DevtoolsErrorType {
-  /**
-   * The name of the error.
-   */
-  name: string
-  /**
-   * How the error is initialized.
-   */
-  initializer: (query: Query) => Error
-}
-
-export interface QueryDevtoolsProps {
-  readonly client: QueryClient
-  queryFlavor: string
-  version: string
-  onlineManager: typeof onlineManager
-
-  buttonPosition?: DevtoolsButtonPosition
-  position?: DevtoolsPosition
-  initialIsOpen?: boolean
-  errorTypes?: Array<DevtoolsErrorType>
-  shadowDOMTarget?: ShadowRoot
-  onClose?: () => void
-  hideDisabledQueries?: boolean
-  theme?: Theme
-}
+export type {
+  DevtoolsButtonPosition,
+  DevtoolsErrorType,
+  DevtoolsPosition,
+  QueryDevtoolsProps,
+  Theme,
+} from './types'
 
 export const QueryDevtoolsContext = createContext<QueryDevtoolsProps>({
   client: undefined as unknown as QueryClient,
